Starts to write on IGraphics objects
The collection of methods in GraphicsToolBox.cpp help to display text
on an IGraphics object. An example can be seen  in the
TimeGraphicsProvider.cpp class where the current time is printed in the
center of the IGraphics object.
On top of that, Google Mock framework is available in the project to
test complex classes or functions (such as the graphical toolbox).