Debian package (instead of tar.gz) and python
CMakelists.txt has been modified to produce a debian package instead of
the tar.gz archive it was producting. All the other scripts have been
adapted to use the debian package. The package will install the python
part of the project using distutils and setup.py.

The python part of the project is just a simple executable script for
the moment.