Install script, french usage, no more deamon
Running this script will create the service that starts automatically
the program on the Pi. This is temporary and should be moved to cpack.
Since this is the way it is meant to be, the python-deamon dependency
and associated code has been removed.

The python program displays the date in french now.